finalize design
